requirements.txt
peewee==3.17.1
psycopg2==2.9.9
requests==2.31.0
deepdiff==6.7.1
python-dotenv==1.0.1